requirements.txt
PySide6
winrt-runtime
winrt-Windows.Media.Control
winrt-Windows.Storage.Streams
winrt-Windows.Foundation
winrt-Windows.Foundation.Collections
